Give each sort its own file, and the list extensions their shape back
IListExtensions.cs held 3,603 lines, of which seventeen sort algorithms were 3,000. Each algorithm and its private helpers now live in Runtime/Core/Extension/Sorting/, the SortAlgorithm enum has its own file, and IListExtensions.cs keeps only the list operations. Pure move: no member body changed.

Refs #430
