+++ Runtime/Core/Extension/Sorting/IListSortBlockMerge.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ccc12523b29f0868a77f547d56eb84d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
