+++ Runtime/Core/Extension/Sorting/IListSortDispatch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d3b2726a9d7de28fa2a7a4b6b789bd11
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
