+++ Runtime/Core/Extension/Sorting/IListSortFlux.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 754881aea8f9d183309623c77ed4572f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
