+++ Runtime/Core/Extension/Sorting/IListSortGhost.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d3d9eb3d01d93482676915897cc30bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
