+++ Runtime/Core/Extension/Sorting/IListSortGlide.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: abdb82d1003cbc539995fc195c1f2d87
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
