+++ Runtime/Core/Extension/Sorting/IListSortInsertion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 596a79fce5514d4db48c5bbe7ef4b187
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
