+++ Runtime/Core/Extension/Sorting/IListSortIpn.cs
@@ -1,0 +1,106 @@
+// MIT License - Copyright (c) 2026 wallstop
+// Full license text: https://github.com/wallstop/unity-helpers/blob/main/LICENSE
+
+// ReSharper disable once CheckNamespace
+namespace WallstopStudios.UnityHelpers.Core.Extension
+{
+    using System.Collections.Generic;
+
+    public static partial class IListExtensions
+    {
+        /// <summary>
+        /// Sorts the list using IpnSort, an unstable introspective quicksort with median-of-medians sampling.
+        /// </summary>
+        /// <remarks>
+        /// Implementation reference: ipnsort (Voultapher) — https://github.com/Voultapher/sort-research-rs/tree/main/writeup/ipnsort_introduction.
+        /// </remarks>
+        public static void IpnSort<T, TComparer>(this IList<T> array, TComparer comparer)
+            where TComparer : IComparer<T>
+        {
+            int count = array.Count;
+            if (count < 2)
+            {
+                return;
+            }
+
+            IpnSortRange(array, 0, count - 1, comparer, 2 * FloorLog2(count));
+        }
+
+        private static void IpnSortRange<T, TComparer>(
+            IList<T> array,
+            int left,
+            int right,
+            TComparer comparer,
+            int depthLimit
+        )
+            where TComparer : IComparer<T>
+        {
+            const int insertionThreshold = 32;
+            while (right - left > insertionThreshold)
+            {
+                if (depthLimit == 0)
+                {
+                    HeapSortRange(array, left, right, comparer);
+                    return;
+                }
+
+                int pivotIndex = IpnSelectPivotIndex(array, left, right, comparer);
+                (int pivotStart, int pivotEnd, bool swapped) = PartitionRange(
+                    array,
+                    left,
+                    right,
+                    pivotIndex,
+                    comparer
+                );
+
+                depthLimit--;
+                if (!swapped && IsRangeSorted(array, left, right, comparer))
+                {
+                    return;
+                }
+
+                if (pivotStart - left < right - pivotEnd)
+                {
+                    if (left < pivotStart)
+                    {
+                        IpnSortRange(array, left, pivotStart - 1, comparer, depthLimit);
+                    }
+                    left = pivotEnd + 1;
+                }
+                else
+                {
+                    if (pivotEnd < right)
+                    {
+                        IpnSortRange(array, pivotEnd + 1, right, comparer, depthLimit);
+                    }
+                    right = pivotStart - 1;
+                }
+            }
+
+            InsertionSortRange(array, left, right, comparer);
+        }
+
+        private static int IpnSelectPivotIndex<T, TComparer>(
+            IList<T> array,
+            int left,
+            int right,
+            TComparer comparer
+        )
+            where TComparer : IComparer<T>
+        {
+            int length = right - left + 1;
+            if (length < 128)
+            {
+                return SelectPivotIndex(array, left, right, comparer);
+            }
+
+            int step = length / 4;
+            int a = left;
+            int b = left + step;
+            int c = left + (step << 1);
+            int d = left + step * 3;
+            int e = right;
+            return MedianOfFiveIndices(array, a, b, c, d, e, comparer);
+        }
+    }
+}
