+++ Runtime/Core/Extension/Sorting/IListSortIpn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ccd52ba3069fb4dde5ea5f9d9c09efd2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
