+++ Runtime/Core/Extension/Sorting/IListSortIps4o.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 636f665533f1777f15af329f45ed2058
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
