+++ Runtime/Core/Extension/Sorting/IListSortMeteor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 381e9ce0b8d84620c22ba1b1616b4583
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
