+++ Runtime/Core/Extension/Sorting/IListSortPatternDefeatingQuick.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 02351a2217f959e66635d3eb8f62d606
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
