+++ Runtime/Core/Extension/Sorting/IListSortPower.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2730bbe38cf3501ad0f0c5f91c1a38f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
