+++ Runtime/Core/Extension/Sorting/IListSortPowerPlus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc4fa31aae6ee6d23d55e71d799ec2ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
