+++ Runtime/Core/Extension/Sorting/IListSortShared.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb59de503e0a7bca64a0399b9f377ce5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
