+++ Runtime/Core/Extension/Sorting/IListSortSka.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48b616647a4d7fc252b8055c8098a5ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
