+++ Runtime/Core/Extension/Sorting/IListSortSmooth.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d8f9bbd75cf038bd42309cb5e5ee878
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
