+++ Runtime/Core/Extension/Sorting/SortAlgorithm.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b8de9eeda11da869fbf3220c5786e54b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
